+++ packages/visual-stack-docs/src/containers/App/SideNav.js
@@ -2,19 +2,20 @@
 /** @prettier */
 import React from 'react';
 import R from 'ramda';
+
 import { Link as RRLink } from 'react-router';
 import LayoutIcon from 'mdi-react/TelevisionGuideIcon';
 import ComponentIcon from 'mdi-react/HexagonMultipleIcon';
 import IconsIcon from 'mdi-react/ShapePlusIcon';
 
 import {
-  SideNav,
   Link,
-  LinkGroup,
   LinkContentWrapper,
+  LinkGroup,
+  SideNav,
+  UserMenu,
 } from '@cjdev/visual-stack-redux/lib/components/SideNav';
 import CJLogo from '@cjdev/visual-stack/lib/components/CJLogo';
-import { UserIcon } from '@cjdev/visual-stack/lib/components/UserMenu';
 import { routeComponentMap } from '../Components/Docs/';
 
 const componentLinks = R.pipe(
@@ -23,33 +24,22 @@ const componentLinks = R.pipe(
   R.sortBy(R.prop('linkName'))
 )(routeComponentMap);
 
-const AppUserMenu = ({ onLogout }) => (
-  <LinkGroup
-    className="sidenav-user-menu"
-    label="Sally Jones"
-    icon={<UserIcon firstInitial="S" lastInitial="J" />}
-  >
-    <Link>
-      <a onClick={onLogout}>
-        <LinkContentWrapper icon={<IconsIcon />} label="Logout" />
-      </a>
-    </Link>
-  </LinkGroup>
-);
-
 export default class AppSideNav extends React.Component {
-  /* s1:start */
-  handleLogout() {
-    alert('handleLogout');
-  }
-  /* s1:end */
-
   render() {
+  /* s1:start */
+    const userMenu =
+      <UserMenu
+        onLogout={() => { alert('handleLogout'); }}
+        label="Victoria Smith"
+        firstInitial="V"
+        lastInitial="S"
+      />;
+  /* s1:end */
     return (
     <div>
     {/* s2:start */}
       <SideNav
-        userMenu={<AppUserMenu onLogout={this.handleLogout} />}
+        userMenu={userMenu}
         initializedCollapsed={false}
         logoBackground="#00AF66"
         logo={<CJLogo />}
