+++ packages/visual-stack-redux/src/components/SideNav/index.js
@@ -3,7 +3,8 @@ export {
   Header,
   Link,
   LinkContentWrapper,
+  LogoutLink,
   SideNavIcon,
 } from '@cjdev/visual-stack/lib/components/SideNav';
 export { LinkGroup } from './LinkGroup';
-export { SideNav } from './SideNav';
+export { SideNav, UserMenu } from './SideNav';
