+++ packages/visual-stack/src/components/SideNav/index.js
@@ -6,3 +6,4 @@ export { LinkGroup } from './LinkGroup';
 export { Header } from './Header';
 export { Link, LinkContentWrapper } from './Link';
 export { SideNav } from './SideNav';
+export { LogoutLink, UserIcon } from './UserMenu';
